backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator
requests
